+++ src/data/talentsData.ts
@@ -1,0 +1,117 @@
+import { CardTalent } from '../interfaces/Cards/CardTalent'
+
+const talentsData: CardTalent[] = [
+  {
+    imageUrl: '/img/talents/placeholder.png',
+    name: 'Miguel Calero',
+    description: 'Magía y hechicería',
+    socialNetworks: [
+      {
+        name: 'Email',
+        url: ''
+      },
+      {
+        name: 'Twitter',
+        url: ''
+      },
+      {
+        name: 'Linkedin',
+        url: ''
+      },
+      {
+        name: 'Dev.to',
+        url: ''
+      }
+    ]
+  },
+  {
+    imageUrl: '/img/talents/placeholder.png',
+    name: 'Javier Aguirre',
+    description: 'Magía y hechicería',
+    socialNetworks: [
+      {
+        name: 'Email',
+        url: ''
+      }
+    ]
+  },
+  {
+    imageUrl: '/img/talents/placeholder.png',
+    name: 'Jesús Mejías',
+    description: 'Magía y hechicería',
+    socialNetworks: [
+      {
+        name: 'Email',
+        url: ''
+      }
+    ]
+  },
+  {
+    imageUrl: '/img/talents/placeholder.png',
+    name: 'Miguel Gavilán',
+    description: 'Magía y hechicería',
+    socialNetworks: [
+      {
+        name: 'Email',
+        url: ''
+      }
+    ]
+  },
+  {
+    imageUrl: '/img/talents/placeholder.png',
+    name: 'Javier Martín',
+    description: 'Magía y hechicería',
+    socialNetworks: [
+      {
+        name: 'Email',
+        url: ''
+      }
+    ]
+  },
+  {
+    imageUrl: '/img/talents/placeholder.png',
+    name: 'Daniel Ranchal',
+    description: 'Magía y hechicería',
+    socialNetworks: [
+      {
+        name: 'Email',
+        url: ''
+      }
+    ]
+  },
+  {
+    imageUrl: '/img/talents/placeholder.png',
+    name: 'Nieves borrego',
+    description: 'Magía y hechicería',
+    socialNetworks: [
+      {
+        name: 'Email',
+        url: ''
+      }
+    ]
+  },
+  {
+    imageUrl: '/img/talents/placeholder.png',
+    name: 'Victoria',
+    description: 'Magía y hechicería',
+    socialNetworks: [
+      {
+        name: 'Email',
+        url: ''
+      }
+    ]
+  },
+  {
+    imageUrl: '/img/talents/placeholder.png',
+    name: 'Maca Torralba',
+    description: 'Magía y hechicería',
+    socialNetworks: [
+      {
+        name: 'Email',
+        url: ''
+      }
+    ]
+  }
+]
+
+export default talentsData