+++ src/interfaces/Cards/CardSimpleBigTitle.d.ts
@@ -1,5 +1,3 @@
-/* eslint-disable camelcase */
-
 export interface CardSimpleBigTitle {
   title: string
   description: string